% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_standardized_scores.R
\name{add_standardized_scores}
\alias{add_standardized_scores}
\title{Add Standardized Scores to Data from REDCap}
\usage{
add_standardized_scores(dat)
}
\arguments{
\item{dat}{\code{tibble} or \code{data.frame} that includes, at the very least, the following:
\itemize{
\item \code{cog_studyid}
\item \code{cog_test_date}
\item \code{cog_sex}
\item \code{cog_education}
\item \code{cog_age}
\item any number of variables that are present in \code{male_female} or \code{otmt} or \code{ravlt_trials_m_sd}
}}
}
\description{
This functions adds a plethora of standardized scores along with some grouping
variables needed to calculate the standardized scores.
}
