% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_plots.R
\name{base_plot_z_scores}
\alias{base_plot_z_scores}
\title{Create base plotly figure}
\usage{
base_plot_z_scores(
  dat,
  cog_vars = c("cog_tmta_time", "cog_otmta_time", "cog_otmta_error", "cog_nsf_total",
    "cog_nsf_span", "cog_nsb_total", "cog_nsb_span"),
  fill_values = setNames(RColorBrewer::brewer.pal(n = 7, "RdYlGn"), nm = c("Impaired",
    "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior")),
  shade_descriptions = T,
  fill_alpha = 0.2,
  source = "A"
)
}
\arguments{
\item{dat}{data object to get ranges for axes from}

\item{cog_vars}{vector of variables to infer ranges from \strong{without prefix}}

\item{fill_values}{a named vector of length seven with hex color values to use for categories
"Impaired", "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior"}

\item{shade_descriptions}{logical; should ranges for descriptions be shaded?}

\item{fill_alpha}{opacity value for shaded areas}

\item{source}{passed to \code{plotly::plot_ly}}
}
\description{
Create a base plot where traces can later be added.
}
\keyword{internal}
