% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_vars_groups.R
\docType{data}
\name{cog_vars_groups}
\alias{cog_vars_groups}
\title{Groups for Cognitive Variables}
\format{
\subsection{\code{cog_vars_groups}}{

\describe{Named vector given groups (in the entries) for cognitive variables (in names).
For example, \code{cog_vars_groups["cog_cdr_global"]} returns the group \code{cog_cdr_global}
belongs to.
}
}
}
\usage{
cog_vars_groups
}
\description{
Groups for Cognitive Variables
}
\keyword{datasets}
