% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_vars_labels.R
\docType{data}
\name{cog_vars_labels}
\alias{cog_vars_labels}
\title{Labels for Cognitive Variables}
\format{
\subsection{\code{cog_vars_labels}}{

\describe{A named vector with labels for the cognitive variables. The names
of the vector are the variable names in the REDCap database, and the entries
of the vector are the corresponding label.
}
}
}
\usage{
cog_vars_labels
}
\description{
Labels for Cognitive Variables
}
\keyword{datasets}
