% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowat_ranges.R
\docType{data}
\name{cowat_ranges}
\alias{cowat_ranges}
\title{COWAT ranges}
\format{
\subsection{\code{cowat_ranges}}{

\describe{Each of the columns AAA, BBB, ..., KKK gives cutoffs that are mapped
to intervals. Each interval corresponds to the standardized score that is in
the row where the beginning of the interval is given. For example, if an
individual from age group BBB has a \code{cog_flc_flu} score of 15, it falls in
the interval [13, 17), and the standardized score is then 3.
}
}
}
\usage{
cowat_ranges
}
\description{
COWAT ranges
}
\keyword{datasets}
