% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepModule.R
\name{dataPrepUI}
\alias{dataPrepUI}
\alias{dataPrepServer}
\title{Shiny Module for Preparing Data from the WLS REDCap Project}
\usage{
dataPrepUI(id)

dataPrepServer(id)
}
\arguments{
\item{id}{An ID string to match module UI and module server.}
}
\value{
A \code{tibble} with the results of \code{prep_data()}.
}
\description{
This shiny module presents options to either pull data straight from REDCap
or upload a .csv file as downloaded from REDCap. To pull from REDCap, a password
is required (contact me \href{mailto:rtrane@wisc.edu}{here}).
}
