% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo Data Set}
\format{
\subsection{\code{demo_data}}{

\describe{This is a completely scrambled version of a subset of WLS participants. \code{ptid/cog_studyid}
are assigned randomly, and all other variables have been randomly reassigned in such a
way that missing values were preserved. This might have caused some abnormalities such
as some "participants" being younger at later visits than they were at earlier visits.
}
}
}
\usage{
demo_data
}
\description{
Demo Data Set
}
\keyword{datasets}
