% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_raw_data.R
\docType{data}
\name{demo_raw_data}
\alias{demo_raw_data}
\title{Demo Raw Data Set}
\format{
\subsection{\code{demo_raw_data}}{

\describe{This is a completely scrambled version of a subset of WLS participants. \code{ptid/cog_studyid}
are assigned randomly, and all other variables have been randomly reassigned in such a
way that missing values were preserved. This might have caused some abnormalities such
as some "participants" being younger at later visits than they were at earlier visits. This data
set does not include standardized scores.
}
}
}
\usage{
demo_raw_data
}
\description{
Demo Raw Data Set
}
\keyword{datasets}
