% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_table.R
\name{demographics_table}
\alias{demographics_table}
\title{Summary Table with Demographic Information}
\usage{
demographics_table(dat, studyid = "102038g")
}
\arguments{
\item{dat}{Data from REDCap}

\item{studyid}{String with ID for participant of interest. Much be length 1.}
}
\description{
Summary Table with Demographic Information
}
\examples{
demographics_table(demo_data, "326231g")

}
