% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digsym_ranges.R
\docType{data}
\name{digsym_ranges}
\alias{digsym_ranges}
\title{WAIS-R Digit Symbols ranges}
\format{
\subsection{\code{digsym_ranges}}{

\describe{Each of the columns AAAA, BBBB, ..., FFFF gives cutoffs that are mapped
to intervals. Each interval corresponds to the standardized score that is in
the row where the beginning of the interval is given. For example, if an
individual from age group BBBB has a \code{cog_digsym} score of 15, it falls in
the interval [14, 18), and the standardized score is then 5.
}
}
}
\usage{
digsym_ranges
}
\description{
WAIS-R Digit Symbols ranges
}
\keyword{datasets}
