% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_gtExtras_for_my_gt_blt_bar_pct.R
\name{ideal_fgnd_color}
\alias{ideal_fgnd_color}
\title{Helpers from gtExtras used in \code{my_gt_blt_bar_pct}}
\usage{
ideal_fgnd_color(bgnd_color, light = "#FFFFFF", dark = "#000000")
}
\description{
Helpers from gtExtras used in \code{my_gt_blt_bar_pct}
}
\keyword{internal}
