% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_measures.R
\name{functional_measures_table}
\alias{functional_measures_table}
\title{Summary Table with Functional Measures}
\usage{
functional_measures_table(dat)
}
\arguments{
\item{dat}{Data set to use; see \code{demo_data} for example}
}
\description{
Summary Table with Functional Measures
}
\examples{
functional_measures_table(dplyr::filter(demo_data, .data$cog_studyid == .data$cog_studyid[1]))

}
