% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_group.R
\name{get_age_group}
\alias{get_age_group}
\title{Get Age Group for Standardized Scores}
\usage{
get_age_group(
  age,
  group_type = c("ravlt_trials", "ravlt_recog", "UDS_battery", "otmt", "cowat", "digsym")
)
}
\arguments{
\item{age}{numeric vector with ages to be converted to groups}

\item{group_type}{string indicator which age group to return. Must be one of
"ravlt_trials", "ravlt_recog", "UDS_battery", "otmt", "cowat", "digsym"}
}
\description{
Get various versions of age groups for finding standardized scores.
}
