% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_descriptions}
\alias{get_descriptions}
\title{Get Description}
\usage{
get_descriptions(raw, z_score, ss, cog_var)
}
\arguments{
\item{raw}{Raw score of cognitive variable}

\item{z_score}{z-scores corresponding to raw scores}

\item{ss}{Standardized scores corresponding to raw scores}

\item{cog_var}{Name of cognitive variable}
}
\description{
Get Description
}
\keyword{internal}
