% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentiles.R
\name{get_percentiles}
\alias{get_percentiles}
\title{Get Percentiles}
\usage{
get_percentiles(z_score, ss, cog_var)
}
\arguments{
\item{z_score}{numeric vector with z-scores}

\item{ss}{numeric vector with standardized scores}

\item{cog_var}{character vector indicating what cognitive score each entry of
\code{z_score} and \code{ss} correspond to.}
}
\description{
Get percentiles from z-scores, Standardized Scores and name of cognitive variable.
}
\keyword{internal}
