% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longTableModule.R
\name{longTableUI}
\alias{longTableUI}
\alias{longTableServer}
\title{Shiny Module to Display Longitudinal Trends in Table}
\usage{
longTableUI(id)

longTableServer(
  id,
  dat,
  fill_values = setNames(RColorBrewer::brewer.pal(n = 7, "RdYlGn"), nm = c("Impaired",
    "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior")),
  table_font_size = 100
)
}
\arguments{
\item{id}{string to tie UI to Server}

\item{dat}{data to use for table. Should only have data on one subject}

\item{fill_values}{a named vector of length seven with hex color values to use for categories
"Impaired", "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior"}

\item{table_font_size}{Percentage font size to use. Used to scale table}
}
\description{
Table version of the plotCogVarModule.
}
