% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_table.R
\name{main_table}
\alias{main_table}
\title{Create NACC T-Cog Neuropsychological Assessment Summary Table}
\usage{
main_table(
  dat,
  fill_values = setNames(RColorBrewer::brewer.pal(n = 7, "RdYlGn"), nm = c("Impaired",
    "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior")),
  bar_height = 16
)
}
\arguments{
\item{dat}{Data to use. Must have exactly one row (data from one participant and one visit).}

\item{fill_values}{a named vector of length seven with hex color values to use for categories
"Impaired", "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior"}

\item{bar_height}{In pixels. Height of the percentile bars. Default: 16}
}
\value{
An object of class \code{gt::gt_tbl}
}
\description{
This function create a summary table of measures from the NACC T-Cog Neuropsychological
Assessment.
}
\examples{
main_table(
  dat = demo_data[1,]
)

}
