% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainTableModule.R
\name{mainTableUI}
\alias{mainTableUI}
\alias{mainTableServer}
\title{Shiny Module to Display NACC T-Cog Neuropsychological Assessment Summary Table}
\usage{
mainTableUI(id)

mainTableServer(id, dat, studyid, date, table_font_size = 100)
}
\arguments{
\item{id}{An ID string to match module UI and server UI}

\item{dat}{Data set to use}

\item{studyid}{Study ID for participant to create summary table for}

\item{date}{The visit date to use for the table}

\item{table_font_size}{Percentage font size to use. Used to scale table}
}
\description{
Shiny Module to Display NACC T-Cog Neuropsychological Assessment Summary Table
}
