% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otmt.R
\docType{data}
\name{otmt}
\alias{otmt}
\title{Oral Trail Making Test: means and standard deviations}
\format{
\subsection{\code{otmt}}{

\describe{
\item{name}{Variable name in WLS REDCap}
\item{otmt_age_group}{Age group as given by Mrazik et al. (2010); see source}
\item{m}{Mean of \code{variable} in \code{otmt_age_group}}
\item{sd}{Standard deviation of \code{variable} in \code{otmt_age_group}}
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2858599/}
}
\usage{
otmt
}
\description{
Averages and standard deviations as reported by Mrazik et al. (2010) for time
to complete and number of errors by age groups. Note: the age groups are
overlapping as follows:
AA: 20-39
BB: 29-49
CC: 39-59
DD: 49-69
EE: 59-79
FF: 69-90
}
\keyword{datasets}
