% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cog_var.R
\name{plot_cog_var}
\alias{plot_cog_var}
\title{Function to Plot Cognitive Variables over Time}
\usage{
plot_cog_var(
  dat,
  var_to_plot = "cog_flc_flu",
  type = c("raw", "standardized", "ss"),
  shade_descriptions = F,
  trim = Inf
)
}
\arguments{
\item{dat}{Data to use}

\item{var_to_plot}{variable to plot \strong{without prefix}}

\item{type}{Type of score to plot. One of "raw", "standardized", or "ss",
depending on \code{var_to_plot}}

\item{shade_descriptions}{logical. Default: FALSE. If TRUE, the figure will be shaded
according to the description column of the main table.}

\item{trim}{default: Inf. Value used to trim scores. Any score with
abs(score) > trim will be removed.}
}
\description{
Function to Plot Cognitive Variables over Time
}
