% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCogVarModule.R
\name{plotCogVarApp}
\alias{plotCogVarApp}
\title{Shiny App Wrapping the variable plotting module}
\usage{
plotCogVarApp(
  dat,
  studyids = c("143050g", "241123g", "301022s", "322002g", "372034g", "382030s",
    "391310s", "440250g", "461231g", "498013s", "524005g", "595004s", "635012g")
)
}
\arguments{
\item{dat}{Data from REDCap to use}

\item{studyids}{vector of studyids to include in drop-down menu}
}
\description{
Shiny App Wrapping the variable plotting module
}
