% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCogVarModule.R
\name{plotCogVarUI}
\alias{plotCogVarUI}
\alias{plotCogVarServer}
\title{Shiny Module to Plot Scores Across Visits}
\usage{
plotCogVarUI(id)

plotCogVarServer(id, redcap_dat, studyid)
}
\arguments{
\item{id}{An ID string to matche module UI and module server}

\item{redcap_dat}{Data to use for the plots. Should contain a number of \verb{cog_*}
variables}

\item{studyid}{The study id of the participant of interest}
}
\description{
This shiny module is meant for visualizing changes over time of a long list
of cognitive measures from the WLS ILIAD study.
}
