% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data from REDCap}
\usage{
prep_data(
  decrypt_password,
  file,
  nas = c("NA", "N/A", "-", " ", "", "#REF!", "17/22", "#DIV/0!", "#VALUE!", "0,5",
    "23a", "28b")
)
}
\arguments{
\item{decrypt_password}{password to decrypt REDCap API Token so data can be
fetched straight from REDCap.}

\item{file}{path to file downloaded from REDCap}

\item{nas}{character vector of strings that should be treated as NA}
}
\description{
This function prepares data downloaded from REDCap to be presented using \code{main_table()}.
It removes records without a valid age before running \code{add_standardized_scores()},
which adds age groups and standardized scores.
}
\examples{
prep_data(
  file = "~/Downloads/WisconsinLongitudina_DATA_2024-08-10_1500.csv",
  nas = c("NA", "N/A", "-", " ", "", "#REF!", "17/22", "#DIV/0!", "#VALUE!")
)

}
