% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravlt_recog_ranges.R
\docType{data}
\name{ravlt_recog_ranges}
\alias{ravlt_recog_ranges}
\title{RAVLT Recognition ranges}
\format{
\subsection{\code{ravlt_recog_ranges}}{

\describe{Each of the columns AAAAA, BBBBB, ..., JJJJJ gives cutoffs that are mapped
to intervals. Each interval corresponds to the standardized score that is in
the row where the beginning of the interval is given. There are separate ranges
for men and women. For example, if an individual from age group EEEEE, who is a male,
has a \code{cog_ravlt_recog_acc} score of 80\%, it falls in
the interval [80, 87), and the standardized score is then 9.
}
}
}
\usage{
ravlt_recog_ranges
}
\description{
RAVLT Recognition ranges
}
\keyword{datasets}
