% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravlt_trials_m_sd.R
\docType{data}
\name{ravlt_trials_m_sd}
\alias{ravlt_trials_m_sd}
\title{RAVLT Means and SDs}
\format{
\subsection{\code{ravlt_trials_m_sd}}{

\describe{
\item{ravlt_age_groups}{Age group as given above}
\item{name}{Variable name as presented in REDCap data}
\item{m}{Average to use for age group and variable}
\item{sd}{SD to use for age group and variable}
}
}
}
\usage{
ravlt_trials_m_sd
}
\description{
Averages and standard deviations for some Rey AVLT outcomes by age groups.
The age groups are as follows:
1: [16-19]
2: [20-29]
3: [30-39]
4: [40-49]
5: [50-59]
6: [60-69]
7: [70-79]
8: >79
}
\keyword{datasets}
