% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_cog_digsym.R
\name{standardized_cog_digsym}
\alias{standardized_cog_digsym}
\title{Get standardized WAIS-R Digit Symbol score}
\usage{
standardized_cog_digsym(cog_digsym, digsym_age_groups)
}
\arguments{
\item{cog_digsym}{Raw score}

\item{digsym_age_groups}{age group; a vector of same length as \code{cog_digsym}
with entries as one of the following
\itemize{
\item 'AAAA' for ages 45-54
\item 'BBBB' for ages 55-64
\item 'CCCC' for ages 65-69
\item 'DDDD' for ages 70-74
\item 'EEEE' for ages 75-79
\item 'FFFF' for ages 80+
}}
}
\value{
A numeric vector of same length as \code{cog_digsym} with the standardized scores.
}
\description{
Get standardized WAIS-R Digit Symbol score
}
\examples{
standardized_cog_digsym(24, 'DDDD')

}
