% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_cog_flc_flu.R
\name{standardized_cog_flc_flu}
\alias{standardized_cog_flc_flu}
\title{Get standardized F+L+C Words score}
\usage{
standardized_cog_flc_flu(cog_flc_flu, cowat_age_groups)
}
\arguments{
\item{cog_flc_flu}{Raw F+l+C score}

\item{cowat_age_groups}{COWAT age group; one of the following
\itemize{
\item 'AAA' for ages 56-62
\item 'BBB' for ages 63-65
\item 'CCC' for ages 66-68
\item 'DDD' for ages 69-71
\item 'EEE' for ages 72-74
\item 'FFF' for ages 75-77
\item 'GGG' for ages 78-80
\item 'HHH' for ages 81-83
\item 'III' for ages 84-86
\item 'JJJ' for ages 87-89
\item 'KKK' for ages 90-97
}}
}
\value{
A numeric vector of same length as \code{cog_flc_flu} with the standardized scores.
}
\description{
This function provides standardized F+L+C scores given the raw score and age group.
}
\examples{
standardized_cog_flc_flu(20, 'HHH')

}
