% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_digsym.R
\name{standardized_digsym}
\alias{standardized_digsym}
\title{Get standardized WAIS-R Digit Symbol score}
\usage{
standardized_digsym(raw_score, age)
}
\arguments{
\item{raw_score}{numeric vector with raw score}

\item{age}{a numeric vector of same length as \code{raw_score} with ages of participants}
}
\value{
A numeric vector of same length as \code{raw_score} with the standardized scores.
}
\description{
Get standardized WAIS-R Digit Symbol score
}
\examples{
standardized_digsym(24, 80)

}
