% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_flc_flu.R
\name{standardized_flc_flu}
\alias{standardized_flc_flu}
\title{Get standardized F+L+C Words score}
\usage{
standardized_flc_flu(raw_score, age)
}
\arguments{
\item{raw_score}{numeric vector}

\item{age}{numeric vector of same length as \code{raw_score} with ages}
}
\value{
A numeric vector of same length as \code{raw_score} with the standardized scores.
}
\description{
This function provides standardized F+L+C scores given the raw score and age group.
}
\examples{
standardized_flc_flu(20, age = 70)

}
