% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_ravlt_recog_acc.R
\name{standardized_ravlt_recog_acc}
\alias{standardized_ravlt_recog_acc}
\title{Get standardized RAVLT Recognition Accuracy score}
\usage{
standardized_ravlt_recog_acc(raw_score, age, sex)
}
\arguments{
\item{raw_score}{Raw score as a percentage (0-100)}

\item{age}{numeric vector of same length as \code{raw_score}}

\item{sex}{vector of same length as \code{raw_score} containing the corresponding
sex of the participants. Each entry must be either "m" (male) or "f" (female)}
}
\value{
A numeric vector of same length as \code{raw_score} with the standardized scores.
}
\description{
Get standardized RAVLT Recognition Accuracy score
}
\examples{
standardized_ravlt_recog_acc(raw_score = 75, age = 60, sex = "m")

}
