% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_labels.R
\name{var_labels}
\alias{var_labels}
\title{Create Row Labels for Main Table}
\usage{
var_labels(
  cog_cdr_sob,
  cog_tmta_err,
  cog_tmta_lines,
  cog_benson_retained,
  cog_benson_recog,
  cog_ravlt_a1,
  cog_ravlt_a2,
  cog_ravlt_a3,
  cog_ravlt_a4,
  cog_ravlt_a5,
  cog_ravlt_recog_tp,
  cog_ravlt_recog_tn,
  cog_tmtb_err,
  cog_tmtb_lines
)
}
\description{
The main table of interest uses some of the observed data in the row labels.
This function creates the row labels as a named vector for easy use in
\code{stringr::str_replace_all} to create a new column with the desired row labels.
All parameters should be "raw" values.
}
