% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsRedcapApp-package.R
\docType{package}
\name{wlsRedcapApp-package}
\alias{wlsRedcapApp-package}
\title{wlsRedcapApp: NACC T-Cog Neuropsychological Assessments}
\description{
Functions for working with NACC T-Cog Neuropsychological Assessments. Most notably this includes a Shiny application specifically for work with WLS REDCap data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmtrane.github.io/wlsRedcapApp/}
}

}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rtrane@wisc.edu}

}
\keyword{internal}
