% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_measures.R
\name{functional_measures_table}
\alias{functional_measures_table}
\title{Summary Table with Functional Measures}
\usage{
functional_measures_table(dat, studyid = "102038g")
}
\arguments{
\item{dat}{Data set to use; see \code{demo_data} for example}

\item{studyid}{String giving participant id. Must be length 1 and one of
the IDs in \code{dat$cog_studyid}.}
}
\description{
Summary Table with Functional Measures
}
\examples{
functional_measures_table(demo_data, demo_data$cog_studyid[1])

}
