% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_table.R
\name{main_table}
\alias{main_table}
\title{Create NACC T-Cog Neuropsychological Assessment Summary Table}
\usage{
main_table(dat, studyid = "102038g", cur_date = "2022-06-13", bar_height = 16)
}
\arguments{
\item{dat}{Data to use}

\item{studyid}{String with ID of a participant. Must be length 1 and present in
\code{dat$cog_studyid}}

\item{cur_date}{String with visit date to use. Must be present in
\code{dplyr::filter(dat, cog_studyid == studyid)$cog_test_date}.}

\item{bar_height}{In pixels. Height of the percentile bars. Default: 16}
}
\value{
An object of class \code{gt::gt_tbl}
}
\description{
This function create a summary table of measures from the NACC T-Cog Neuropsychological
Assessment.
}
\examples{
main_table(
  dat = demo_data,
  studyid = demo_data$cog_studyid[1],
  cur_date = demo_data$cog_test_date[1]
)

}
