% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cog_var.R
\name{plot_cog_var}
\alias{plot_cog_var}
\title{Function to Plot Cognitive Variables over Time}
\usage{
plot_cog_var(
  redcap_dat,
  studyid = "102038g",
  var_to_plot = "cog_flc_flu",
  type = c("raw", "standardized", "ss")
)
}
\arguments{
\item{redcap_dat}{Data to use}

\item{studyid}{ID of participant to plot}

\item{var_to_plot}{variable to plot \strong{without prefix}}

\item{type}{Type of score to plot. One of "raw", "standardized", or "ss",
depending on \code{var_to_plot}}
}
\description{
Function to Plot Cognitive Variables over Time
}
