% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCogVarModule.R
\name{plotCogVarApp}
\alias{plotCogVarApp}
\title{Shiny App Wrapping the variable plotting module}
\usage{
plotCogVarApp(dat, trim = Inf)
}
\arguments{
\item{dat}{Data from REDCap to use}

\item{trim}{default: Inf. Value used to trim scores. Any score with
abs(score) > trim will be removed.}
}
\description{
Shiny App Wrapping the variable plotting module
}
