% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCogVarModule.R
\name{plotCogVarApp}
\alias{plotCogVarApp}
\title{Shiny App Wrapping the variable plotting module}
\usage{
plotCogVarApp(dat, studyid = "102038g")
}
\arguments{
\item{dat}{Data from REDCap to use}

\item{studyid}{The id of the participant of interest}
}
\description{
Shiny App Wrapping the variable plotting module
}
