% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCogVarModule.R
\name{plotCogVarUI}
\alias{plotCogVarUI}
\alias{plotCogVarServer}
\title{Shiny Module to Plot Scores Across Visits}
\usage{
plotCogVarUI(id)

plotCogVarServer(
  id,
  redcap_data,
  studyid,
  fill_values = setNames(RColorBrewer::brewer.pal(n = 7, "RdYlGn"), nm = c("Impaired",
    "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior"))
)
}
\arguments{
\item{id}{An ID string to match module UI and module server}

\item{redcap_data}{Data to use for the plots. Should contain a number of \verb{cog_*}
variables}

\item{studyid}{ID of participant; reactive}

\item{fill_values}{a named vector of length seven with hex color values to use for categories
"Impaired", "Borderline", "Low Average", "Average", "High Average", "Superior", "Very Superior"}
}
\description{
This shiny module is meant for visualizing changes over time of a long list
of cognitive measures from the WLS ILIAD study.
}
\examples{
\dontrun{
  plotCogVarApp(demo_data[1,])
}


}
