% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_cog_ravlt_recog_acc.R
\name{standardized_cog_ravlt_recog_acc}
\alias{standardized_cog_ravlt_recog_acc}
\title{Get standardized RAVLT Recognition Accuracy score}
\usage{
standardized_cog_ravlt_recog_acc(
  cog_ravlt_recog_acc,
  ravlt_recog_age_groups,
  sex
)
}
\arguments{
\item{cog_ravlt_recog_acc}{Raw score as a percentage (0-100)}

\item{ravlt_recog_age_groups}{age group; a vector of same length as \code{cog_digsym}
with entries as one of the following
\itemize{
\item 'AAAAA' for ages [60,62]
\item 'BBBBB' for ages [63,65]
\item 'CCCCC' for ages [66,68]
\item 'DDDDD' for ages [69,71]
\item 'EEEEE' for ages [72,74]
\item 'FFFFF' for ages [75,77]
\item 'GGGGG' for ages [78,80]
\item 'HHHHH' for ages [81,83]
\item 'IIIII' for ages [84,86]
\item 'JJJJJ' for ages >87
}}

\item{sex}{vector of same length as \code{cog_digsym} containing the corresponding
sex of the participants. Each entry must be either "m" (male) or "f" (female)}
}
\value{
A numeric vector of same length as \code{cog_ravlt_recog_acc} with the standardized scores.
}
\description{
Get standardized RAVLT Recognition Accuracy score
}
\examples{
standardized_cog_ravlt_recog_acc(75, 'DDDDD', "m")

}
