% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_labels.R
\name{var_labels}
\alias{var_labels}
\title{Create Row Labels for Main Table}
\usage{
var_labels(
  cog_cdr_sob = NA,
  cog_tmta_err = NA,
  cog_tmta_lines = NA,
  cog_benson_retained = NA,
  cog_benson_recog = NA,
  cog_ravlt_a1 = NA,
  cog_ravlt_a2 = NA,
  cog_ravlt_a3 = NA,
  cog_ravlt_a4 = NA,
  cog_ravlt_a5 = NA,
  cog_ravlt_recog_tp = NA,
  cog_ravlt_recog_tn = NA,
  cog_tmtb_err = NA,
  cog_tmtb_lines = NA,
  cog_craft_verb_retain = NA,
  cog_craft_par_retain = NA
)
}
\arguments{
\item{cog_cdr_sob}{Raw observed value of CDR Sob}

\item{cog_tmta_err}{number of errors made on Trail Making Test A}

\item{cog_tmta_lines}{lines completed on Trail Making Test A}

\item{cog_benson_retained}{percent retained on Benson}

\item{cog_benson_recog}{Benson recognized}

\item{cog_ravlt_a1}{Rey AVLT Trial 1}

\item{cog_ravlt_a2}{Rey AVLT Trial 2}

\item{cog_ravlt_a3}{Rey AVLT Trial 3}

\item{cog_ravlt_a4}{Rey AVLT Trial 4}

\item{cog_ravlt_a5}{Rey AVLT Trial 5}

\item{cog_ravlt_recog_tp}{Rey AVLT Recognition True Positives}

\item{cog_ravlt_recog_tn}{Rey AVLT Recognition True Negatives}

\item{cog_tmtb_err}{number of errors made on Trail Making Test B}

\item{cog_tmtb_lines}{lines completed on Trail Making Test B}

\item{cog_craft_verb_retain}{percentage retained from cog_craft_imm_verb to cog_craft_delay_verb}

\item{cog_craft_par_retain}{percentage retained from cog_craft_imm_par to cog_craft_delay_par}
}
\description{
The main table of interest uses some of the observed data in the row labels.
This function creates the row labels as a named vector for easy use in
\code{stringr::str_replace_all} to create a new column with the desired row labels.
All parameters should be "raw" values.
}
