% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsRedcapApp-package.R
\docType{package}
\name{wlsRedcapApp-package}
\alias{wlsRedcapApp-package}
\title{wlsRedcapApp: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmtrane.github.io/wlsRedcapApp/}
}

}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rtrane@wisc.edu}

}
\keyword{internal}
